+++ R/RcppExports.R
@@ -41,6 +41,14 @@ descendant_tips <- function(parent, child, postorder) {
     .Call(`_TreeTools_descendant_tips`, parent, child, postorder)
 }
 
+first_matching_split_pair <- function(x, table) {
+    .Call(`_TreeTools_first_matching_split_pair`, x, table)
+}
+
+first_matching_split_index <- function(x, table) {
+    .Call(`_TreeTools_first_matching_split_index`, x, table)
+}
+
 num_to_parent <- function(n, nTip) {
     .Call(`_TreeTools_num_to_parent`, n, nTip)
 }