+++ tests/testthat/test-ClusterTable.R
@@ -114,6 +114,7 @@ test_that("ClusterTable with multiple trees", {
 
 test_that("ClusterTable with complex trees", {
   skip_if_not_installed("TreeDist", "2.9.2.9000")
+  skip_on_cran()
   library("TreeDist")
   
   # Test exposes failures in C++ - constexpr not playing nicely with Rcpp